+++ packages/core/src/components/essentials/tooltip-axis.ts
@@ -1,5 +1,9 @@
 import { Tooltip } from "./tooltip";
-import { AxisPositions, ScaleTypes } from "../../interfaces";
+import {
+	AxisPositions,
+	ScaleTypes,
+	ColorClassNameTypes
+} from "../../interfaces";
 import { Tools } from "../../tools";
 
 import { format } from "date-fns";
@@ -77,7 +81,11 @@ export class AxisChartsTooltip extends Tooltip {
 				{
 					label: options.tooltip.groupLabel || "Group",
 					value: datum[groupMapsTo],
-					color: this.model.getStrokeColor(datum[groupMapsTo])
+					color: this.model.getFillColor(datum[groupMapsTo]),
+					class: this.model.getColorClassName({
+						classNameTypes: [ColorClassNameTypes.TOOLTIP],
+						dataGroupName: datum[groupMapsTo]
+					})
 				}
 			];
 		} else if (data.length > 1) {
@@ -93,7 +101,11 @@ export class AxisChartsTooltip extends Tooltip {
 					.map((datum) => ({
 						label: datum[groupMapsTo],
 						value: this.valueFormatter(datum[cartesianScales.getRangeIdentifier(datum)]),
-						color: this.model.getStrokeColor(datum[groupMapsTo])
+						color: this.model.getFillColor(datum[groupMapsTo]),
+						class: this.model.getColorClassName({
+							classNameTypes: [ColorClassNameTypes.TOOLTIP],
+							dataGroupName: datum[groupMapsTo]
+						})
 					}))
 					.sort((a, b) => b.value - a.value)
 			);