+++ packages/core/src/configuration.ts
@@ -158,6 +158,10 @@ const chart: BaseChartOptions = {
 	},
 	color: {
 		scale: null,
+		pairing: {
+			numberOfVariants: null,
+			option: 1
+		},
 		gradient: {
 			enabled: false
 		}