+++ packages/core/src/model.ts
@@ -1,8 +1,7 @@
 // Internal Imports
 import * as Configuration from "./configuration";
 import { Tools } from "./tools";
-import * as colorPalettes from "./services/colorPalettes";
-import { Events, ScaleTypes } from "./interfaces";
+import { Events, ScaleTypes, ColorClassNameTypes } from "./interfaces";
 
 // D3
 import { map } from "d3-collection";
@@ -34,6 +33,8 @@ export class ChartModel {
 
 	// Fill scales & fill related objects
 	protected colorScale: any = {};
+
+	protected colorClassNames: any = {};
 
 	constructor(services: any) {
 		this.services = services;
@@ -359,7 +360,8 @@ export class ChartModel {
 
 		this.updateAllDataGroups();
 
-		this.setColorScale();
+		this.setCustomColorScale();
+		this.setColorClassNames();
 		this.services.events.dispatchEvent(Events.Model.UPDATE, { animate });
 	}
 
@@ -458,8 +460,12 @@ export class ChartModel {
 	}
 
 	getFillColor(group: any, key?: any, data?: any) {
+		if (!this.isUserProvidedColorScaleValid()) {
+			return null;
+		}
 		const options = this.getOptions();
 		const defaultFillColor = this.getFillScale()(group);
+
 		if (options.getFillColor) {
 			return options.getFillColor(group, key, data, defaultFillColor);
 		} else {
@@ -468,6 +474,10 @@ export class ChartModel {
 	}
 
 	getStrokeColor(group: any, key?: any, data?: any) {
+		if (!this.isUserProvidedColorScaleValid()) {
+			return null;
+		}
+
 		const options = this.getOptions();
 		const defaultStrokeColor = this.colorScale(group);
 		if (options.getStrokeColor) {
@@ -479,6 +489,46 @@ export class ChartModel {
 
 	getFillScale() {
 		return this.colorScale;
+	}
+
+	isUserProvidedColorScaleValid() {
+		const userProvidedScale = Tools.getProperty(
+			this.getOptions(),
+			"color",
+			"scale"
+		);
+		const dataGroups = this.getDataGroups();
+
+		if (
+			userProvidedScale == null ||
+			Object.keys(userProvidedScale).length < dataGroups.length
+		) {
+			return false;
+		}
+
+		return dataGroups.every((dataGroup) =>
+			Object.keys(userProvidedScale).includes(dataGroup.name)
+		);
+	}
+
+	getColorClassName(configs: {
+		classNameTypes: ColorClassNameTypes[];
+		dataGroupName: string;
+		originalClassName?: string;
+	}) {
+		if (this.isUserProvidedColorScaleValid()) {
+			return configs.originalClassName;
+		}
+
+		const colorPairingTag = this.colorClassNames(configs.dataGroupName);
+		let className = configs.originalClassName;
+		configs.classNameTypes.forEach(
+			(type) =>
+				(className = configs.originalClassName
+					? `${className} ${type}-${colorPairingTag}`
+					: `${type}-${colorPairingTag}`)
+		);
+		return className;
 	}
 
 	/**
@@ -611,47 +661,81 @@ export class ChartModel {
 	/*
 	 * Fill scales
 	 */
-	protected setColorScale() {
-		const defaultColors = colorPalettes.DEFAULT;
+	protected setCustomColorScale() {
+		if (!this.isUserProvidedColorScaleValid()) {
+			return;
+		}
 
 		const options = this.getOptions();
 		const userProvidedScale = Tools.getProperty(options, "color", "scale");
 
-		// If there is no valid user provided scale, use the default set of colors
-		if (
-			userProvidedScale === null ||
-			Object.keys(userProvidedScale).length === 0
-		) {
-			this.colorScale = scaleOrdinal()
-				.range(defaultColors)
-				.domain(this.allDataGroups);
-
-			return;
-		}
-
 		/**
 		 * Go through allDataGroups. If a data group has a color value provided
 		 * by the user, add that to the color range
-		 * If not, add a default color
 		 */
 		const colorRange = [];
-		let colorIndex = 0;
 		this.allDataGroups.forEach((dataGroup) => {
 			if (userProvidedScale[dataGroup]) {
 				colorRange.push(userProvidedScale[dataGroup]);
-			} else {
-				colorRange.push(defaultColors[colorIndex]);
-			}
-
-			if (colorIndex === defaultColors.length - 1) {
-				colorIndex = 0;
-			} else {
-				colorIndex++;
 			}
 		});
 
 		this.colorScale = scaleOrdinal()
 			.range(colorRange)
 			.domain(this.allDataGroups);
+	}
+
+	/*
+	 * Color palette
+	 */
+	protected setColorClassNames() {
+		const colorPairingOptions = Tools.getProperty(
+			this.getOptions(),
+			"color",
+			"pairing"
+		);
+
+		// Check if user has defined numberOfVariants (differ from given data)
+		let numberOfVariants = Tools.getProperty(
+			colorPairingOptions,
+			"numberOfVariants"
+		);
+		if (!numberOfVariants || numberOfVariants < this.allDataGroups.length) {
+			numberOfVariants = this.allDataGroups.length;
+		}
+
+		let pairingOption = Tools.getProperty(colorPairingOptions, "option");
+		const colorPairingCounts = {
+			"1-color": 4,
+			"2-color": 5,
+			"3-color": 5,
+			"4-color": 3,
+			"5-color": 2,
+			"14-color": 1
+		};
+
+		// If number of dataGroups is greater than 5, user 14-color palette
+		const numberOfColors = numberOfVariants > 5 ? 14 : numberOfVariants;
+
+		// Use default palette if user choice is not in range
+		pairingOption =
+			pairingOption <= colorPairingCounts[`${numberOfColors}-color`]
+				? pairingOption
+				: 1;
+
+		// Create color classes for graph, tooltip and stroke use
+		const colorPairing = this.allDataGroups.map(
+			(dataGroup, index) =>
+				`${numberOfColors}-${pairingOption}-${(index % 14) + 1}`
+		);
+
+		// If there is no valid user provided scale, use the default set of colors
+		if (!this.isUserProvidedColorScaleValid()) {
+			this.colorClassNames = scaleOrdinal()
+				.range(colorPairing)
+				.domain(this.allDataGroups);
+
+			return;
+		}
 	}
 }